requirements.txt
# Deep Learning Framework
# PyTorch 2.9.0 with CUDA 12.6 support
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
# For CUDA 12.6, install with:
# pip install torch==2.9.0 torchvision==0.24.0 torchaudio==2.9.0 --index-url https://download.pytorch.org/whl/cu126

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Financial Data
yfinance>=0.2.0

# Technical Analysis
pandas-ta>=0.3.14b
# ta-lib requires system-level installation, install separately if needed

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Backtesting
backtrader>=1.9.78

# Experiment Tracking
mlflow>=2.8.0
tensorboard>=2.14.0

# API Development
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
click>=8.1.0

# Data Storage
pyarrow>=13.0.0
h5py>=3.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# GPU Support (Optional)
# For NVIDIA GPUs (CUDA)
# torch with CUDA support will be installed based on system
# For AMD GPUs (ROCm)
# torch-rocm can be installed separately if needed

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0